+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added read to `cartesi-rollups-cli` for inputs, notices, vouchers and reports
 - Added `Makefile` to auto-generate project files
 - Added experimental sunodo validator mode
+- Added inspect command to `cartesi-rollups-cli`
 
 ### Changed